Moved Point to a proper class and documented it
