+++ include/Header.h
@@ -4,7 +4,7 @@
 #include <fstream>
 #include <memory>
 
-#include "ezc3d.h"
+#include <ezc3d.h>
 
 class EZC3D_API ezc3d::Header{
 public:
